initfs: update: Don't exec sh or sulogin on error just exit 1
When update was written it was exec'd from the shutdown script
and hence took over pid 1.  Since exiting in that environment was
a panic situation, the script instead started a rescue shell with
its output presumably on the console.

The calling convention was updated to be a simple call to the shell
script in dbacf104885c3bf78c5f3e06f63bba6db0742e4b but the error
handling was not updated.  That error handling is now becoming
a hinderance to use from additional environments so change it.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>