initfs: packaging: Create a package to tell init to run from RAM
Adding the obmc-phosphor-initfs-ram package will add the text
overlay-filesystem-in-ram to /init-options-base which will cause
init to not mount the rwfs.  This allows building an image for
booting over the network that will completely run from RAM,
while allowing copy-files-to-ram to the kernel command line via
bootargs will allow whitelisted files read from the rwfs before
startup and allows them to be saved back to the rwfs via update.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>