[RFC] Add ram boot option to default obmc-phosphor-image recipe
This adds the obmc-phosphor-initfs-ram package which packages the
initramfs-options-base to run from RAM.

This is really more of an example than suggestion, and was created
while exploring packaging options.  If we were to run with this
option, no updates would be written to the flash, except possibly
by random invocations of the update script.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>