initfs: systemd v254 initrd interface workarounds
Bind mount /run/initramfs ourselves before mounting ro and rw
Unmount the tree under /run/initramfs (but not /run/initramfs)
Unmount /ro and /rw - were under /run/initramfs before the pivot

systemd added a soft-reboot feature in v254 and changed the shutdown
unmount semantics to preserving the mount upon umount error.  In
addition for some unknown reason the automatic bind mount of
/run/initramfs is broken for some reason; at lest there is an invalid
argument when trying to manipulate the mount.

While leaving the underlying mounts is useful in that we don't loose
track that the storage is actually still in use by the kernel mount,
it turns out to be very difficult to find the run/initramfs/ro and rw
mount underneath the bind mounted run/initramfs pivoted to /.  Instead
its much easier to perform the mouunt where we need it.

Signed-off-by: Milton Miller <mdmii@outlook.com>