initfs shutdown: adapt to unmounted /run/initramfs
Bind mount / to run/initramfs where it was durinig system operation
if run/initramfs/shutdown is missing

Between v254 and v254.1 systemd decided not to mount /run/initramfs
recursively, with the assumption it was unused since its available
at /, but this means that the update script and image files to be
applied are not available at the runtime location.  However we can
restore this mount ourselves.

Adopt and use the api variable used in init

Signed-off-by: Milton Miller <mdmii@outlook.com>