initfs: Relax final watchdog timeout after update
Bump the watchdog timeout from 15s to 150s aka 2.5m after update
has completed, relaxing the reaction time before reboot.  The
normal path still calls reboot -f to reboot when the console has
been flushed.

When transfering control to our initrd shutdown, systemd leaves
a watchdog running to make sure the system doesn't hang.  The
update was determined to be too short to complete updating some
spi flashes, leaving them with erased but not written images in
the spi-nor mtd, effectively bricking the machines.  So to keep it
alive we start a watchdogd daemon to ping it every 1s timeout 5s
during flash update.  However after flash update completes we were
setting the timeout to 15s leaving it armed.  While this is enough
time to flush the serial console and call reboot, its pretty tight
for a human to recognise a the shell prompt and type the echo -n V
to the watchdog device to disarm it.

Signed-off-by: Milton Miller <mdmii@outlook.com>