initfs shutdown: relax final watchdog timeout
Bump the watchdog reboot from 15s to 150s aka 2.5m after update

We start a watchdog keepalive daemon to ping every 1s timeout 5s
during flash update.  However after flash update completes we were
setting the timeout to 15s leaving it armed.  While this is enough
time to flush the serial console and call reboot, its pretty tight
for a human to recognise a the shell prompt and type the echo -n V
to the watchdog device to disarm it.

Signed-off-by: Milton Miller <mdmii@outlook.com>