initfs: Restore our former scripts
Revert to our original initramfs init, shutdown, and update
scripts to test with systemd top of tree.

The former mount and unmount strategy used by phosphor-initfs
has been in use for close to 6 years.  Revert to it to provide
clarity on what has changed when systemd hands off and the
effects it has on our real world shutdown and update process.

Use the experience from the prior results to guide the structure
for discussion.

Signed-off-by: Milton Miller <mdmii@outlook.com>