initfs: Mount ro and rw filesystems under /run/mnt
Experiment with ro and rw under /run/mnt instead of the api instead of
under the API directory /run/initramfs to avoid confusion with the bind
mount systemd does to pivot_root into shutdown.

Explictly create the api directory, leave commented out lines to revert
and or mount either a tmpfs or just the bind mount ourselves for simple
editing.

Teach update to use the new location if the mount exists.

Signed-off-by: Milton Miller <mdmii@outlook.com>