initfs: Setup run/initramfs after options
Defer setting up the INITRD_INTERFACE api directory with the copy
for shutdown until after the options files are assembled.  This
allows us to use our standard init options to choose tmpfs, bind,
or no mount before populating.

If the options are under the mount we need to preserve them before
mounting tmpfs.

Always copy the urlfile if it exists in case consider_download is
added to options on a later reboot.

(I planed for but did not test storing the options outside of api).

Signed-off-by: Milton Miller <mdmii@outlook.com>