initfs: Prepare shutdown for experiments
Structure code to experiment with different strategies to deal
with the systemd behaviour changes from v253.3 to v254 and v254.1

Log (to console) the mounts and init-options file at invocation.

Save the to-unmount filter to a file, with commented lines to
adjust what is filtered: The traditional oldmount and mnt, adding
/run/initramfs bind mount, or just the directories underneath
/run/initramfs, and finally matching /ro and /rw (but not explictly
the / because I don't believe busybox awk is POSIX compliant).

If /run/initramfs/shutdown is missing then we mounted a tmpfs and
systemd v254.1 did not bind it back onto /run/initramfs so do that
now.

Note: the comments might refer to commented behavior and not all
combinations are valid for all combination of bind and tmpfs mounts.

Signed-off-by: Milton Miller <mdmii@outlook.com>